Engendrando un tema... (para reservas pero capaz que lo usamos)

+++ trunk/includes/classes/Admin/ProyectoActividad/Helper.php
@@ -7,7 +7,7 @@ class Admin_ProyectoActividad_Helper extends Core_Singleton{
 		if(!is_a($proyecto_actividad,'Inta_Model_ProyectoActividad')){
 			$proyecto_actividad = new Inta_Model_ProyectoActividad($proyecto_actividad->getData());
 		}
-		$existente = $proyecto_actividad->setWhere(Db_Helper::equal('id_actividad'),Db_Helper::equal('id_proyecto'))->searchCount();
+		$existente = $proyecto_actividad->setWhere(Db_Helper::equal('id_actividad'),Db_Helper::equal('id_proyecto'))->search();
 		if(!$proyecto_actividad->hasId()){/** aca hay que agregar a la base de datos*/
 			if($existente){
 				Admin_App::getInstance()->addInfoMessage("La ProyectoActividad ya estaba agregada");
@@ -28,7 +28,7 @@ class Admin_ProyectoActividad_Helper extends Core_Singleton{
 		}
 		else{/** aca hay que actualizar el registro*/
 			//$actualizada = true;// actualizarEnLaBase()
-			if($existente){
+			if($existente&&$existente[0]->getId()!=$proyecto_actividad->getId()){
 				$proyecto_actividad->delete(array('id'=>$proyecto_actividad->getId()));
 				Admin_App::getInstance()->addInfoMessage("La ProyectoActividad ya estaba agregada");
 				$resultado = true;
