+++ trunk/includes/classes/Admin/UsuarioActividad/Helper.php
@@ -7,7 +7,7 @@ class Admin_UsuarioActividad_Helper extends Core_Singleton{
 		if(!is_a($usuario_actividad,'Inta_Model_UsuarioActividad')){
 			$usuario_actividad = new Inta_Model_UsuarioActividad($usuario_actividad->getData());
 		}
-		$existente = $usuario_actividad->setWhere(Db_Helper::equal('id_actividad'),Db_Helper::equal('id_usuario'))->searchCount();
+		$existente = $usuario_actividad->setWhere(Db_Helper::equal('id_actividad'),Db_Helper::equal('id_usuario'))->search();
 		if(!$usuario_actividad->hasId()){/** aca hay que agregar a la base de datos*/
 			if($existente){
 				Admin_App::getInstance()->addInfoMessage("El UsuarioActividad ya estaba agregado");
@@ -27,7 +27,7 @@ class Admin_UsuarioActividad_Helper extends Core_Singleton{
 		}
 		else{/** aca hay que actualizar el registro*/
 			//$actualizada = true;// actualizarEnLaBase()
-			if($existente){
+			if($existente&&$existente[0]->getId()!=$usuario_actividad->getId()){
 				$usuario_actividad->delete(array('id'=>$usuario_actividad->getId()));
 				Admin_App::getInstance()->addInfoMessage("El UsuarioActividad ya estaba agregado");
 				$resultado = true;
